WritableUtils: fix VLong handling
Python's WritableUtils wasn't lining up with Java's Writable;
for example, the integer 12195 is encoded as "\x8e\x2f\xa3"
with the java version, but "\x8e\x2f\xbb" with the Python version.

This appears to be a bug with the integer math when the high bit
is set on a byte; fix by reading/writing unsigned characters instead.