Added way for users to create Writables and specify the Java names to save in the SequenceFile, overriding the reflection-based determination.
hadoop.util.ReflectionUtils.hadoopClassName now checks for the existence of hadoop_module_name and hadoop_class_name class attributes before pulling the hadoop class name from the Python module and class name.

This is useful because it lets you subclass AbstractValueWritable and implement other Hadoop (such as BytesWritable in my case) writables in your code, rather than needing to modify the python-hadoop code. This also provides an easy way to implement arbitrary Writables that are not part of Hadoop, provided an equivalent Java class is used with Hadoop jobs.